update from com.github.iruzhnikov to io.github.iruzhnikov (sonatype issue)
