+++ spring-boot/spring-webmvc-cors-properties-autoconfigure/pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>com.github.iruzhnikov</groupId>
+        <groupId>io.github.iruzhnikov</groupId>
         <artifactId>spring-boot-cors-properties</artifactId>
         <version>0.0.1</version>
         <relativePath>../pom.xml</relativePath>
@@ -15,7 +15,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.github.iruzhnikov</groupId>
+            <groupId>io.github.iruzhnikov</groupId>
             <artifactId>spring-webmvc-cors-properties</artifactId>
         </dependency>
         <dependency>
