--- spring-boot/spring-webmvc-cors-properties-autoconfigure/src/main/java/com/github/iruzhnikov/web/config/YamlPropertyLoaderFactory.java
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.iruzhnikov.web.config;
-
-import org.springframework.boot.env.YamlPropertySourceLoader;
-import org.springframework.core.env.PropertySource;
-import org.springframework.core.io.support.DefaultPropertySourceFactory;
-import org.springframework.core.io.support.EncodedResource;
-import org.springframework.lang.Nullable;
-
-import java.io.IOException;
-import java.util.List;
-
-/**
- * Check PropertySource support or
- * StandardConfigDataLocationResolver#configNames check add custom names
- */
-public class YamlPropertyLoaderFactory extends DefaultPropertySourceFactory {
-    @Override
-    public PropertySource<?> createPropertySource(@Nullable String name, EncodedResource resource) throws IOException {
-        List<PropertySource<?>> propertySources =
-                new YamlPropertySourceLoader().load(resource.getResource().getFilename(), resource.getResource());
-        return propertySources.get(0);
-    }
-}
