+++ spring-boot/spring-webmvc-cors-properties-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.github.iruzhnikov.web.config.CorsPropWebMvcConfiguration,\
-  com.github.iruzhnikov.web.servlet.CorsPropWebMvcConfigurationSupport,\
-  com.github.iruzhnikov.web.config.DefaultCorsMvcConfiguration
+  io.github.iruzhnikov.web.config.CorsPropWebMvcConfiguration,\
+  io.github.iruzhnikov.web.servlet.CorsPropWebMvcConfigurationSupport,\
+  io.github.iruzhnikov.web.config.DefaultCorsMvcConfiguration
