+++ spring-boot/spring-webmvc-cors-properties-autoconfigure/src/test/java/io/github/iruzhnikov/TestApplication.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.iruzhnikov;
+
+import io.github.iruzhnikov.web.servlet.CorsPropWebMvcConfigurationSupport;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.AutoConfigurationPackage;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+
+
+@EnableWebSecurity
+@AutoConfigurationPackage(basePackageClasses = CorsPropWebMvcConfigurationSupport.class)
+@SpringBootApplication
+@ConfigurationPropertiesScan
+@ComponentScan(lazyInit = true)
+public class TestApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TestApplication.class, args);
+    }
+}
