+++ spring-framework/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.github.iruzhnikov</groupId>
+        <groupId>io.github.iruzhnikov</groupId>
         <artifactId>spring-cors-properties</artifactId>
         <version>0.0.1</version>
         <relativePath>../pom.xml</relativePath>
@@ -24,7 +24,7 @@
     <dependencyManagement>
         <dependencies>
             <dependency>
-                <groupId>com.github.iruzhnikov</groupId>
+                <groupId>io.github.iruzhnikov</groupId>
                 <artifactId>spring-webmvc-cors-properties</artifactId>
                 <version>${self.version}</version>
             </dependency>
