+++ spring-framework/spring-webmvc-cors-properties/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.github.iruzhnikov</groupId>
+        <groupId>io.github.iruzhnikov</groupId>
         <artifactId>spring-framework-cors-properties</artifactId>
         <version>0.0.1</version>
         <relativePath>../pom.xml</relativePath>
