--- spring-framework/spring-webmvc-cors-properties/src/main/java/com/github/iruzhnikov/web/config/PathCorsConfiguration.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.iruzhnikov.web.config;
-
-import lombok.Data;
-
-import javax.annotation.Nullable;
-import java.time.Duration;
-
-@Data
-public class PathCorsConfiguration {
-    //Default pattern
-    public static final String ALL_PATH = "/**";
-
-    private String path;
-    @Nullable
-    private String[] allowedOrigins;
-    @Nullable
-    private String[] allowedOriginPatterns;
-    @Nullable
-    private String[] allowedMethods;
-    @Nullable
-    private String[] allowedHeaders;
-    @Nullable
-    private String[] exposedHeaders;
-    @Nullable
-    private Boolean allowCredentials;
-    @Nullable
-    private Duration maxAge;
-}
